content/html/content/test/test_bug763626.html

<!DOCTYPE html>
<html>



<head>
<title>Test for Bug 763626</title>

<script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

<script type="application/javascript">

SimpleTest.waitForExplicitFinish();

function boom()
{
  var r = document.createElement("div").itemRef;
  SpecialPowers.DOMWindowUtils.garbageCollect();
  is("" + r, "", "ToString should return empty string when element is gone");
  SimpleTest.finish();
}

</script>
</head>

<body onload="boom();"></body>
</html>
